# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Janauber API",
    "description": "API for Managing janauber"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "API-KEY",
      "description": "API JWT token"
    }
  },
  "consumes": [
    "application/json"
  ],
  "paths": {
    "/passenger/{id}": {
      "get": {
        "summary": "GET one user",
        "description": "This endpoint is responsible to get ONE users.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/passenger/": {
      "get": {
        "summary": "GET users",
        "description": "This endpoint is responsible to get ALL users.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "summary": "POST user",
        "description": "This endpoint is responsible to create a new user.",
        "parameters": [
          {
            "name": "passenger",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            },
            "description": "Create New Passenger"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "cpf": {
                    "example": "any"
                  },
                  "birthday": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "autoBody": true
      }
    },
    "/account/login": {
      "post": {
        "summary": "POST Log in a User (can be a driver ou a passenger )",
        "description": "This endpoint is responsible to Login ALL users.",
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            },
            "description": "Login an User"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/account/register/passenger": {
      "post": {
        "summary": "POST Register Passenger",
        "description": "This endpoint is responsible to register a new Passenger.",
        "parameters": [
          {
            "name": "register",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            },
            "description": "Login an User"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "cpf": {
                    "example": "any"
                  },
                  "birthday": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/driver/{id}": {
      "get": {
        "summary": "GET one driver",
        "description": "This endpoint is responsible to get ONE driver.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/driver/": {
      "get": {
        "summary": "GET drivers",
        "description": "This endpoint is responsible to gel all drivers",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "post": {
        "summary": "POST driver",
        "description": "This endpoint is responsible to create a new driver",
        "parameters": [
          {
            "name": "driver",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "cpf": {
                  "example": "any"
                },
                "birthday": {
                  "example": "any"
                }
              }
            },
            "description": "Create New Driver"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "cpf": {
                    "example": "any"
                  },
                  "birthday": {
                    "example": "any"
                  }
                }
              }
            }
          }
        },
        "autoBody": true
      }
    }
  }
}